seperate out calibration script into modules
