+++ unpackaged/vesuvio_calibration/tests/unit/test_calibrate_vesuvio_fit.py
@@ -1,9 +1,6 @@
-import sys
-
-from unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5 import EVSCalibrationFit, DETECTOR_RANGE, \
-     ENERGY_ESTIMATE, BRAGG_PEAK_CROP_RANGE, BRAGG_FIT_WINDOW_RANGE, RECOIL_PEAK_CROP_RANGE, RECOIL_FIT_WINDOW_RANGE, \
-     RESONANCE_PEAK_CROP_RANGE, RESONANCE_FIT_WINDOW_RANGE, PEAK_HEIGHT_RELATIVE_THRESHOLD
-from mock import MagicMock, patch, call, ANY
+from unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit import EVSCalibrationFit
+from unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_helper_functions import EVSGlobals
+from mock import MagicMock, patch, call
 from functools import partial
 from mantid.kernel import IntArrayProperty, StringArrayProperty, FloatArrayProperty
 
@@ -33,9 +30,9 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
     def side_effect_cell(row_index, col_index, peaks):
         return peaks[row_index]
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.DeleteWorkspace')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.CloneWorkspace')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.mtd')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.DeleteWorkspace')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.CloneWorkspace')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.mtd')
     def test_filter_peaks_perfect_match(self, mock_mtd, mock_clone_workspace, mock_del_workspace):
         alg = EVSCalibrationFit()
 
@@ -67,9 +64,9 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
         mock_del_workspace.assert_called_with(find_peaks_output_name + '_unfiltered')
 
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.DeleteWorkspace')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.CloneWorkspace')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.mtd')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.DeleteWorkspace')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.CloneWorkspace')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.mtd')
     def test_filter_peaks_no_match(self, mock_mtd, mock_clone_workspace, mock_del_workspace):
         alg = EVSCalibrationFit()
 
@@ -99,9 +96,9 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
                                                 OutputWorkspace=find_peaks_output_name + '_unfiltered')
         mock_del_workspace.assert_called_with(find_peaks_output_name + '_unfiltered')
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.DeleteWorkspace')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.CloneWorkspace')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.mtd')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.DeleteWorkspace')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.CloneWorkspace')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.mtd')
     def test_filter_peaks_one_match(self, mock_mtd, mock_clone_workspace, mock_del_workspace):
         alg = EVSCalibrationFit()
 
@@ -132,9 +129,9 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
                                                 OutputWorkspace=find_peaks_output_name + '_unfiltered')
         mock_del_workspace.assert_called_with(find_peaks_output_name + '_unfiltered')
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.DeleteWorkspace')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.CloneWorkspace')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.mtd')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.DeleteWorkspace')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.CloneWorkspace')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.mtd')
     def test_filter_peaks_two_match(self, mock_mtd, mock_clone_workspace, mock_del_workspace):
         alg = EVSCalibrationFit()
 
@@ -166,9 +163,9 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
         mock_del_workspace.assert_called_with(find_peaks_output_name + '_unfiltered')
 
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.DeleteWorkspace')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.CloneWorkspace')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.mtd')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.DeleteWorkspace')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.CloneWorkspace')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.mtd')
     def test_filter_peaks_does_not_include_higher_found_peak(self, mock_mtd, mock_clone_workspace, mock_del_workspace):
         alg = EVSCalibrationFit()
 
@@ -199,9 +196,9 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
                                                 OutputWorkspace=find_peaks_output_name + '_unfiltered')
         mock_del_workspace.assert_called_with(find_peaks_output_name + '_unfiltered')
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.DeleteWorkspace')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.CloneWorkspace')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.mtd')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.DeleteWorkspace')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.CloneWorkspace')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.mtd')
     def test_filter_peaks_does_not_include_lower_found_peak(self, mock_mtd, mock_clone_workspace, mock_del_workspace):
         alg = EVSCalibrationFit()
 
@@ -233,9 +230,9 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
         mock_del_workspace.assert_called_with(find_peaks_output_name + '_unfiltered')
 
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.DeleteWorkspace')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.CloneWorkspace')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.mtd')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.DeleteWorkspace')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.CloneWorkspace')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.mtd')
     def test_filter_peaks_handles_multiple_peaks(self, mock_mtd, mock_clone_workspace, mock_del_workspace):
         alg = EVSCalibrationFit()
 
@@ -266,9 +263,9 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
         mock_del_workspace.assert_called_with(find_peaks_output_name + '_unfiltered')
 
     #Found peaks sometimes returns 'zero' peaks, usually at the end of the table workspace.
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.DeleteWorkspace')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.CloneWorkspace')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.mtd')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.DeleteWorkspace')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.CloneWorkspace')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.mtd')
     def test_filter_peaks_handles_zero_position_in_found_peaks(self, mock_mtd, mock_clone_workspace, mock_del_workspace):
         alg = EVSCalibrationFit()
 
@@ -319,7 +316,7 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
         np.testing.assert_almost_equal([9629.84, 13619.43, 15727.03], estimated_positions.flatten().tolist(), 0.01)
         print(estimated_positions)
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.mtd')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.mtd')
     def test_check_nans_false(self, mock_mtd):
         alg = EVSCalibrationFit()
         table_ws = 'table_ws'
@@ -334,7 +331,7 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
 
         self.assertFalse(alg._check_nans(table_ws))
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.mtd')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.mtd')
     def test_check_nans_true(self, mock_mtd):
         alg = EVSCalibrationFit()
         table_ws = 'table_ws'
@@ -352,9 +349,10 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
     def test_PyInit_property_defaults(self):
         alg = EVSCalibrationFit()
         alg.PyInit()
-        properties = {'Samples': [], 'Background': [], 'Mode': 'FoilOut', 'Function': 'Gaussian', 'SpectrumRange': DETECTOR_RANGE,
-                      'Mass': 207.19, 'DSpacings': [], 'Energy': [ENERGY_ESTIMATE], 'InstrumentParameterFile': '',
-                      'PeakType': '', 'InstrumentParameterWorkspace': None, 'CreateOutput': False, 'OutputWorkspace': ''}
+        properties = {'Samples': [], 'Background': [], 'Mode': 'FoilOut', 'Function': 'Gaussian',
+                      'SpectrumRange': EVSGlobals.DETECTOR_RANGE, 'Mass': 207.19, 'DSpacings': [], 'Energy': [EVSGlobals.ENERGY_ESTIMATE],
+                      'InstrumentParameterFile': '', 'PeakType': '', 'InstrumentParameterWorkspace': None, 'CreateOutput': False,
+                      'OutputWorkspace': ''}
         for prop in properties:
             expected_value = alg.getProperty(prop).value
             if type(expected_value) == np.ndarray:
@@ -362,12 +360,12 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
             self.assertEqual(expected_value, properties[prop], f'Property {prop}. Expected: {expected_value},'
                                                                f'Actual: {properties[prop]}')
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.EVSCalibrationFit._setup_spectra_list')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.EVSCalibrationFit._setup_run_numbers_and_output_workspace')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.EVSCalibrationFit._setup_function_type')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.EVSCalibrationFit._setup_parameter_workspace')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.EVSCalibrationFit._setup_peaks_and_set_crop_and_fit_ranges')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.EVSCalibrationFit._setup_class_variables_from_properties')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.EVSCalibrationFit._setup_spectra_list')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.EVSCalibrationFit._setup_run_numbers_and_output_workspace')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.EVSCalibrationFit._setup_function_type')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.EVSCalibrationFit._setup_parameter_workspace')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.EVSCalibrationFit._setup_peaks_and_set_crop_and_fit_ranges')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.EVSCalibrationFit._setup_class_variables_from_properties')
     def test_setup_calls_all_functions(self, mock_setup_vars, mock_setup_peaks, mock_setup_param_ws, mock_setup_fn_type,
                                         mock_setup_run_nos, mock_setup_spec):
         alg = EVSCalibrationFit()
@@ -482,7 +480,7 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
         self.assertEqual('test_ws', alg._param_workspace)
         self.assertEqual('test_ws', alg._param_table)
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.load_instrument_parameters')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.EVSMiscFunctions.load_instrument_parameters')
     def test_setup_parameter_workspace_no_ws(self, mock_load_instrument_parameters):
         alg = EVSCalibrationFit()
         alg.declareProperty('InstrumentParameterWorkspace', '')
@@ -498,8 +496,8 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
         alg.declareProperty('PeakType', 'Bragg')
         alg._setup_peaks_and_set_crop_and_fit_ranges()
         self.assertEqual(sorted(test_d_spacings), list(alg._d_spacings))
-        self.assertEqual(BRAGG_PEAK_CROP_RANGE, alg._ws_crop_range)
-        self.assertEqual(BRAGG_FIT_WINDOW_RANGE, alg._fit_window_range)
+        self.assertEqual(EVSGlobals.BRAGG_PEAK_CROP_RANGE, alg._ws_crop_range)
+        self.assertEqual(EVSGlobals.BRAGG_FIT_WINDOW_RANGE, alg._fit_window_range)
 
     def test_setup_peaks_and_set_crop_and_fit_ranges_recoil(self):
         test_d_spacings = []
@@ -507,8 +505,8 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
         alg.declareProperty(FloatArrayProperty('DSpacings', test_d_spacings))
         alg.declareProperty('PeakType', 'Recoil')
         alg._setup_peaks_and_set_crop_and_fit_ranges()
-        self.assertEqual(RECOIL_PEAK_CROP_RANGE, alg._ws_crop_range)
-        self.assertEqual(RECOIL_FIT_WINDOW_RANGE, alg._fit_window_range)
+        self.assertEqual(EVSGlobals.RECOIL_PEAK_CROP_RANGE, alg._ws_crop_range)
+        self.assertEqual(EVSGlobals.RECOIL_FIT_WINDOW_RANGE, alg._fit_window_range)
 
     def test_setup_peaks_and_set_crop_and_fit_ranges_resonance(self):
         test_d_spacings = []
@@ -516,11 +514,11 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
         alg.declareProperty(FloatArrayProperty('DSpacings', test_d_spacings))
         alg.declareProperty('PeakType', 'Resonance')
         alg._setup_peaks_and_set_crop_and_fit_ranges()
-        self.assertEqual(RESONANCE_PEAK_CROP_RANGE, alg._ws_crop_range)
-        self.assertEqual(RESONANCE_FIT_WINDOW_RANGE, alg._fit_window_range)
+        self.assertEqual(EVSGlobals.RESONANCE_PEAK_CROP_RANGE, alg._ws_crop_range)
+        self.assertEqual(EVSGlobals.RESONANCE_FIT_WINDOW_RANGE, alg._fit_window_range)
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.ReplaceSpecialValues')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.EVSCalibrationFit._load_to_ads_and_crop')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.ReplaceSpecialValues')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.EVSCalibrationFit._load_to_ads_and_crop')
     def test_preprocess_no_bg(self, mock_load_to_ads_and_crop, mock_replace_special_values):
         test_run_numbers = [1, 2, 3, 4]
         test_crop_range = [3, 10]
@@ -537,9 +535,9 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
         mock_replace_special_values.assert_called_once_with(test_sample_ws_name, NaNValue=0, NaNError=0, InfinityValue=0,
                                                             InfinityError=0, OutputWorkspace=test_sample_ws_name)
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.EVSCalibrationFit._normalise_sample_by_background')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.ReplaceSpecialValues')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.EVSCalibrationFit._load_to_ads_and_crop')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.EVSCalibrationFit._normalise_sample_by_background')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.ReplaceSpecialValues')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.EVSCalibrationFit._load_to_ads_and_crop')
     def test_preprocess_with_bg(self, mock_load_to_ads_and_crop, mock_replace_special_values, mock_normalise_sample):
         test_run_numbers = [1, 2, 3, 4]
         test_bg_run_numbers = [5, 6]
@@ -561,8 +559,8 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
         mock_replace_special_values.assert_called_once_with(test_sample_ws_name, NaNValue=0, NaNError=0, InfinityValue=0,
                                                             InfinityError=0, OutputWorkspace=test_sample_ws_name)
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.CropWorkspace')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.EVSCalibrationFit._load_files')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.CropWorkspace')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.EVSCalibrationFit._load_files')
     def test_load_to_ads_and_crop(self, mock_load_files, mock_crop_workspace):
         alg = EVSCalibrationFit()
         run_numbers = [1, 2, 3, 4]
@@ -574,9 +572,9 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
         mock_load_files.assert_called_once_with(run_numbers, output)
         mock_crop_workspace.assert_called_once_with(output, XMin=xmin, XMax=xmax, OutputWorkspace=output)
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.DeleteWorkspace')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.Divide')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.RebinToWorkspace')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.DeleteWorkspace')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.Divide')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.RebinToWorkspace')
     def test_normalise_sample_by_background(self, mock_rebin, mock_divide, mock_delete):
         alg = EVSCalibrationFit()
         sample_ws = 'test_ws'
@@ -590,9 +588,9 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
         mock_divide.assert_called_once_with(sample_ws, bg_ws, OutputWorkspace=sample_ws)
         mock_delete.assert_called_once_with(bg_ws)
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.DeleteWorkspace')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.Plus')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.EVSCalibrationFit._load_file')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.DeleteWorkspace')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.Plus')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.EVSCalibrationFit._load_file')
     def test_load_files(self, mock_load_file, mock_plus, mock_delete):
         alg = EVSCalibrationFit()
         ws_numbers = ['1-4']  # Note this is parsed as '1-3', is this intentional?
@@ -604,8 +602,8 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
                                     call(output_name, '__EVS_calib_temp_ws', OutputWorkspace=output_name)])
         mock_delete.assert_has_calls([call('__EVS_calib_temp_ws'), call('__EVS_calib_temp_ws')])
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.LoadRaw')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.LoadVesuvio')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.LoadRaw')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.LoadVesuvio')
     def test_load_file_vesuvio(self, mock_load_vesuvio, mock_load_raw):
         alg = EVSCalibrationFit()
         ws_name = 'test_file'
@@ -621,9 +619,9 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
                                                   EnableLogging=False)
         mock_load_raw.assert_not_called()
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.ConvertToDistribution')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.LoadRaw')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.LoadVesuvio')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.ConvertToDistribution')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.LoadRaw')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.LoadVesuvio')
     def test_load_file_raw(self, mock_load_vesuvio, mock_load_raw, mock_convert_to_dist):
         alg = EVSCalibrationFit()
         ws_name = 'test_file'
@@ -640,7 +638,7 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
                                               EnableLogging=False)
         mock_convert_to_dist.assert_called_once_with(output_name, EnableLogging=False)
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.GroupWorkspaces')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.GroupWorkspaces')
     def test_fit_peaks_individual(self, group_workspaces_mock):
         alg = EVSCalibrationFit()
         alg._estimate_peak_positions = MagicMock(return_value=np.asarray([[5, 10, 15], [2.5, 7.5, 10.5]]))
@@ -662,8 +660,8 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
         group_workspaces_mock.assert_called_once_with(['output_ws_name_Peak_0_Parameters', 'output_ws_name_Peak_1_Parameters'],
                                                       OutputWorkspace='output_ws_name_Peak_Parameters')
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.EVSCalibrationFit._shared_parameter_fit')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.GroupWorkspaces')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.EVSCalibrationFit._shared_parameter_fit')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.GroupWorkspaces')
     def test_fit_peaks_shared(self, group_workspaces_mock, shared_parameter_fit_mock):
         alg = EVSCalibrationFit()
         alg._estimate_peak_positions = MagicMock(return_value=np.asarray([[5, 10, 15], [2.5, 7.5, 10.5]]))
@@ -686,8 +684,8 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
                                                       OutputWorkspace='output_ws_name_Peak_Parameters')
         shared_parameter_fit_mock.assert_called_once_with('output_ws_name_Peak_1_Parameters', ['a', 'b', 'c'])
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.EVSCalibrationFit._shared_parameter_fit')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.GroupWorkspaces')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.EVSCalibrationFit._shared_parameter_fit')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.GroupWorkspaces')
     def test_fit_peaks_both(self, group_workspaces_mock, shared_parameter_fit_mock):
         alg = EVSCalibrationFit()
         alg._estimate_peak_positions = MagicMock(return_value=np.asarray([[5, 10, 15], [2.5, 7.5, 10.5]]))
@@ -725,7 +723,7 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
 
         return alg
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.Fit')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.Fit')
     def test_fit_peak(self, mock_fit):
         alg = self._setup_alg_mocks_fit_peak()
 
@@ -749,9 +747,9 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
 
         self.assertEqual(fit_workspace_name, 'fws')
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.mtd')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.DeleteWorkspace')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.FindPeaks')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.mtd')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.DeleteWorkspace')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.FindPeaks')
     def test_find_peaks_and_output_params(self, find_peaks_mock, delete_workspace_mock, mtd_mock):
         alg = EVSCalibrationFit()
         alg._sample = 'sample'
@@ -769,11 +767,11 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
         peak_table_name_ws.rowCount.assert_called_once()
         delete_workspace_mock.assert_called_once_with('__sample_peaks_table_0_3')
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.sys')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.logger.error')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.mtd')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.FindPeaks')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.EVSCalibrationFit._get_find_peak_parameters')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.sys')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.logger.error')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.mtd')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.FindPeaks')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.EVSCalibrationFit._get_find_peak_parameters')
     def test_find_peaks_and_output_params_no_peaks_found(self, find_peak_params_mock, find_peaks_mock, mtd_mock, logger_mock,
                                                          sys_mock):
         alg = EVSCalibrationFit()
@@ -804,7 +802,7 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
         self.assertEqual(3, xmin)
         self.assertEqual(7, xmax)
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.logger.warning')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.logger.warning')
     def test_find_fit_x_window_position_less_than_1(self, logger_mock):
         alg = EVSCalibrationFit()
         alg._func_param_names = {'Position': 'Position_key'}
@@ -825,7 +823,7 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
         else:
             raise ValueError("incorrect column index supplied")
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.mtd')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.mtd')
     def test_output_fit_params_to_table_ws(self, mtd_mock):
         alg = EVSCalibrationFit()
         spec_num = 45
@@ -840,22 +838,22 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
         mtd_mock.__getitem__.assert_called_once_with(output_table_name)
         output_table_ws_mock.addRow.assert_called_once_with([spec_num, 1, 0.1, 2, 0.2, 3, 0.3])
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.DeleteWorkspace')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.DeleteWorkspace')
     def test_del_fit_workspace(self, del_ws_mock):
         alg = EVSCalibrationFit()
         alg._create_output = True
         alg._del_fit_workspaces('ncm', 'params', 'fws')
         del_ws_mock.assert_has_calls([call('ncm'), call('params')])
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.DeleteWorkspace')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.DeleteWorkspace')
     def test_del_fit_workspace_create_output_true(self, del_ws_mock):
         alg = EVSCalibrationFit()
         alg._create_output = False
         alg._del_fit_workspaces('ncm', 'params', 'fws')
         del_ws_mock.assert_has_calls([call('ncm'), call('params'), call('fws')])
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.CreateEmptyTableWorkspace')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.AnalysisDataService')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.CreateEmptyTableWorkspace')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.AnalysisDataService')
     def test_create_output_parameters_table_ws(self, mock_ADS, mock_create_empty_table_ws):
         output_table_name = 'test_output_table'
         num_estimated_peaks = 3
@@ -922,7 +920,7 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
         self.assertEqual(selected_params, fit_results['params'])
         self.assertFalse(unconstrained_fit_selected)
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.mtd')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.mtd')
     def test_output_params_to_table(self, mock_mtd_module):
         alg = EVSCalibrationFit()
         spec_num = 1
@@ -951,7 +949,7 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
         expected_row = [1, 0.1, 2, 0.2, 3, 0.3, 4, 0.4, 5, 0.5]
         mock_output_table.addRow.assert_called_once_with([spec_num] + expected_row)
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.DeleteWorkspace')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.DeleteWorkspace')
     def test_get_output_and_clean_workspaces_unconstrained_not_performed(self, mock_delete_ws):
         alg = EVSCalibrationFit()
         find_peaks_output_name = 'test_find_output_name'
@@ -962,7 +960,7 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
                                          call(find_peaks_output_name)])
         self.assertEqual(output_ws, fit_peaks_output_name + '_Workspace')
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.DeleteWorkspace')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.DeleteWorkspace')
     def test_get_output_and_clean_workspaces_unconstrained_performed(self, mock_delete_ws):
         alg = EVSCalibrationFit()
         find_peaks_output_name = 'test_find_output_name'
@@ -977,7 +975,7 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
                                         call(fit_peaks_output_name + '_unconstrained_Workspace')])
         self.assertEqual(output_ws, fit_peaks_output_name + '_Workspace')
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.DeleteWorkspace')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.DeleteWorkspace')
     def test_get_output_and_clean_workspaces_unconstrained_performed_and_selected(self, mock_delete_ws):
         alg = EVSCalibrationFit()
         find_peaks_output_name = 'test_find_output_name'
@@ -1028,8 +1026,8 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
         mtd_mock_obj.__getitem__.side_effect = lambda name: mock_find_peaks_output if\
             name == find_peaks_name else None
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.FindPeaks')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.mtd')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.FindPeaks')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.mtd')
     def test_run_find_peaks_peaks_found(self, mock_mtd_module, mock_find_peaks):
         alg, fn_args = self._setup_run_find_peaks_test(unconstrained=False)
 
@@ -1039,8 +1037,8 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
         mock_find_peaks.assert_called_once()
         self.assertTrue(result)
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.FindPeaks')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.mtd')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.FindPeaks')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.mtd')
     def test_run_find_peaks_no_peaks_found_raises_value_error(self, mock_mtd_module, mock_find_peaks):
         alg, fn_args = self._setup_run_find_peaks_test(unconstrained=False)
 
@@ -1050,8 +1048,8 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
             alg._run_find_peaks(**fn_args)
         mock_find_peaks.assert_called_once()
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.FindPeaks')
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.mtd')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.FindPeaks')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.mtd')
     def test_run_find_peaks_unconstrained_no_peaks_found_no_error(self, mock_mtd_module, mock_find_peaks):
         alg, fn_args = self._setup_run_find_peaks_test(unconstrained=True)
 
@@ -1061,7 +1059,7 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
         mock_find_peaks.assert_called_once()
         self.assertFalse(result)
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.FindPeaks')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.FindPeaks')
     def test_run_find_peaks_unconstrained_peaks_found_raises_error(self, mock_find_peaks):
         alg, fn_args = self._setup_run_find_peaks_test(unconstrained=True)
 
@@ -1097,12 +1095,12 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
         alg._calc_linear_bg_coefficients.assert_called_once()
         alg._filter_found_peaks.assert_called_once_with(find_peaks_output_name, peak_estimates_list,
                                                              'linear_bg_coeffs',
-                                                             peak_height_rel_threshold=PEAK_HEIGHT_RELATIVE_THRESHOLD)
+                                                             peak_height_rel_threshold=EVSGlobals.PEAK_HEIGHT_RELATIVE_THRESHOLD)
         alg._fit_found_peaks.assert_called_once_with(find_peaks_output_name, None, workspace_index,
                                                           fit_peaks_output_name, x_range)
         alg._check_fitted_peak_validity.assert_called_once_with(fit_peaks_output_name + '_Parameters',
                                                                      peak_estimates_list,
-                                                                     peak_height_rel_threshold=PEAK_HEIGHT_RELATIVE_THRESHOLD)
+                                                                     peak_height_rel_threshold=EVSGlobals.PEAK_HEIGHT_RELATIVE_THRESHOLD)
         self.assertEqual(fit_results['status'], 'test_status')
 
 
@@ -1111,15 +1109,15 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
             x_range = self._setup_filter_and_fit_found_peaks_mocks()
 
         fit_results = alg._filter_and_fit_found_peaks(workspace_index, peak_estimates_list, find_peaks_output_name,
-                                                           fit_peaks_output_name, x_range, unconstrained=False)
+                                                      fit_peaks_output_name, x_range, unconstrained=False)
 
         alg._calc_linear_bg_coefficients.assert_not_called()
         alg._filter_found_peaks.assert_not_called()
         alg._fit_found_peaks.assert_called_once_with(find_peaks_output_name, peak_estimates_list, workspace_index,
-                                                          fit_peaks_output_name, x_range)
+                                                     fit_peaks_output_name, x_range)
         alg._check_fitted_peak_validity.assert_called_once_with(fit_peaks_output_name + '_Parameters',
-                                                                     peak_estimates_list,
-                                                                     peak_height_rel_threshold=PEAK_HEIGHT_RELATIVE_THRESHOLD)
+                                                                peak_estimates_list,
+                                                                peak_height_rel_threshold=EVSGlobals.PEAK_HEIGHT_RELATIVE_THRESHOLD)
         self.assertEqual(fit_results['status'], 'test_status')
 
     def test_filter_and_fit_found_peaks_invalid_peaks(self):
@@ -1128,14 +1126,15 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
         alg._check_fitted_peak_validity.return_value = False
 
         fit_results = alg._filter_and_fit_found_peaks(workspace_index, peak_estimates_list, find_peaks_output_name,
-                                                           fit_peaks_output_name, x_range, unconstrained=False)
+                                                      fit_peaks_output_name, x_range, unconstrained=False)
 
         alg._calc_linear_bg_coefficients.assert_not_called()
         alg._filter_found_peaks.assert_not_called()
         alg._fit_found_peaks.assert_called_once_with(find_peaks_output_name, peak_estimates_list, workspace_index,
                                                           fit_peaks_output_name, x_range)
         alg._check_fitted_peak_validity.assert_called_once_with(fit_peaks_output_name + '_Parameters',
-                                                                     peak_estimates_list, peak_height_rel_threshold=PEAK_HEIGHT_RELATIVE_THRESHOLD)
+                                                                peak_estimates_list,
+                                                                peak_height_rel_threshold=EVSGlobals.PEAK_HEIGHT_RELATIVE_THRESHOLD)
         self.assertEqual(fit_results['status'], 'peaks invalid')
 
     @staticmethod
@@ -1222,7 +1221,7 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
                 except ValueError:
                     np.testing.assert_array_equal(arg, expected_arg)
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.GroupWorkspaces')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.GroupWorkspaces')
     def test_fit_bragg_peaks_success(self, group_workspaces_mock):
         fit_result_ret_val = {'status': 'success'}
         fit_results = lambda *args: fit_result_ret_val
@@ -1244,7 +1243,7 @@ class TestVesuvioCalibrationFit(unittest.TestCase):
                                                                call(False, False, False, 'sample_peaks_table_2', 'output_Spec_2')])
         group_workspaces_mock.assert_called_once_with(','.join(output_workspaces), OutputWorkspace='output_Peak_Fits')
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.GroupWorkspaces')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_fit.GroupWorkspaces')
     def test_fit_bragg_peaks_not_success(self, group_workspaces_mock):
         x_range = (1, 2)
         fit_res = {'status': 'failure', 'xmin': x_range[0], 'xmax': x_range[1]}
