+++ unpackaged/vesuvio_calibration/tests/unit/test_calibrate_vesuvio_misc.py
@@ -1,4 +1,4 @@
-from unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5 import generate_fit_function_header, identify_invalid_spectra
+from unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_helper_functions import EVSMiscFunctions
 from mock import MagicMock, patch
 
 import unittest
@@ -14,27 +14,27 @@ class TestVesuvioCalibrationMisc(unittest.TestCase):
         pass
 
     def test_generate_header_function_gaussian(self):
-        header = generate_fit_function_header("Gaussian")
+        header = EVSMiscFunctions.generate_fit_function_header("Gaussian")
         self.assertEqual({'Height': 'Height', 'Width': 'Sigma', 'Position': 'PeakCentre'}, header)
 
     def test_generate_header_function_gaussian_with_error(self):
-        header = generate_fit_function_header("Gaussian", error=True)
+        header = EVSMiscFunctions.generate_fit_function_header("Gaussian", error=True)
         self.assertEqual({'Height': 'Height_Err', 'Width': 'Sigma_Err', 'Position': 'PeakCentre_Err'}, header)
 
     def test_generate_header_function_voigt(self):
-        header = generate_fit_function_header("Voigt")
+        header = EVSMiscFunctions.generate_fit_function_header("Voigt")
         self.assertEqual({'Height': 'LorentzAmp', 'Position': 'LorentzPos', 'Width': 'LorentzFWHM', 'Width_2': 'GaussianFWHM'}, header)
 
     def test_generate_header_function_voigt_with_error(self):
-        header = generate_fit_function_header("Voigt", error=True)
+        header = EVSMiscFunctions.generate_fit_function_header("Voigt", error=True)
         self.assertEqual({'Height': 'LorentzAmpErr', 'Position': 'LorentzPosErr', 'Width': 'LorentzFWHMErr', 'Width_2': 'GaussianFWHMErr'},
                          header)
 
     def test_generate_header_function_invalid(self):
         with self.assertRaises(ValueError):
-            generate_fit_function_header("Invalid")
+            EVSMiscFunctions.generate_fit_function_header("Invalid")
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.mtd')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_helper_functions.mtd')
     def test_identify_invalid_spectra_no_invalid(self, mock_mtd):
         ws_mock = MagicMock()
         mock_column_output_dict = {'f1.GaussianFWHM': [1, 2, 3], 'f1.GaussianFWHM_Err': [0.1, 0.2, 0.3],
@@ -43,9 +43,10 @@ class TestVesuvioCalibrationMisc(unittest.TestCase):
         ws_mock.column.side_effect = lambda key: mock_column_output_dict[key]
         mock_mtd.__getitem__.return_value = ws_mock
 
-        np.testing.assert_equal(np.argwhere([]), identify_invalid_spectra('peak_table', [5, 10, 20], [0.1, 0.15, 0.2], [0, 2]))
+        np.testing.assert_equal(np.argwhere([]),
+                                EVSMiscFunctions.identify_invalid_spectra('peak_table', [5, 10, 20], [0.1, 0.15, 0.2], [0, 2]))
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.mtd')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_helper_functions.mtd')
     def test_identify_invalid_spectra_nan_in_errors(self, mock_mtd):
         ws_mock = MagicMock()
         mock_column_output_dict = {'f1.GaussianFWHM': [1, 2, 3], 'f1.GaussianFWHM_Err': [np.nan, 0.2, 0.3],
@@ -55,9 +56,9 @@ class TestVesuvioCalibrationMisc(unittest.TestCase):
         mock_mtd.__getitem__.return_value = ws_mock
 
         np.testing.assert_equal(np.argwhere(np.array([True, False, True])),
-                                identify_invalid_spectra('peak_table', [5, 10, 20], [0.1, 0.15, 0.2], [0, 2]))
+                                EVSMiscFunctions.identify_invalid_spectra('peak_table', [5, 10, 20], [0.1, 0.15, 0.2], [0, 2]))
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.mtd')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_helper_functions.mtd')
     def test_identify_invalid_spectra_inf_in_errors(self, mock_mtd):
         ws_mock = MagicMock()
         mock_column_output_dict = {'f1.GaussianFWHM': [1, 2, 3], 'f1.GaussianFWHM_Err': [0.1, 0.2, 0.3],
@@ -67,9 +68,9 @@ class TestVesuvioCalibrationMisc(unittest.TestCase):
         mock_mtd.__getitem__.return_value = ws_mock
 
         np.testing.assert_equal(np.argwhere(np.array([True, True, False])),
-                                identify_invalid_spectra('peak_table', [5, 10, 20], [0.1, 0.15, 0.2], [0, 2]))
+                                EVSMiscFunctions.identify_invalid_spectra('peak_table', [5, 10, 20], [0.1, 0.15, 0.2], [0, 2]))
 
-    @patch('unpackaged.vesuvio_calibration.calibrate_vesuvio_uranium_martyn_MK5.mtd')
+    @patch('unpackaged.vesuvio_calibration.calibration_scripts.calibrate_vesuvio_helper_functions.mtd')
     def test_identify_invalid_spectra_error_greater_than_peak(self, mock_mtd):
         ws_mock = MagicMock()
         mock_column_output_dict = {'f1.GaussianFWHM': [1, 2, 3], 'f1.GaussianFWHM_Err': [0.1, 0.2, 0.3],
@@ -79,7 +80,7 @@ class TestVesuvioCalibrationMisc(unittest.TestCase):
         mock_mtd.__getitem__.return_value = ws_mock
 
         np.testing.assert_equal(np.argwhere(np.array([False, True, True])),
-                                identify_invalid_spectra('peak_table', [5, 10, 20], [0.1, 0.15, 0.2], [0, 2]))
+                                EVSMiscFunctions.identify_invalid_spectra('peak_table', [5, 10, 20], [0.1, 0.15, 0.2], [0, 2]))
 
 
 if __name__ == '__main__':
